@@ -56,44 +54,57 @@
   ],
   "author": "Evgeny Poberezkin",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ajv-validator/ajv/issues"
-  },
-  "homepage": "https://github.com/ajv-validator/ajv",
-  "tonicExampleFilename": ".tonic_example.js",
+  "bugs": "https://github.com/ajv-validator/ajv/issues",
+  "homepage": "https://ajv.js.org",
+  "runkitExampleFilename": ".runkit_example.js",
   "dependencies": {
     "fast-deep-equal": "^3.1.1",
-    "fast-json-stable-stringify": "^2.0.0",
-    "json-schema-traverse": "^0.4.1",
+    "json-schema-traverse": "^1.0.0",
+    "require-from-string": "^2.0.2",
     "uri-js": "^4.2.2"
   },
   "devDependencies": {
-    "ajv-async": "^1.0.0",
-    "bluebird": "^3.5.3",
-    "brfs": "^2.0.0",
-    "browserify": "^16.2.0",
+    "@ajv-validator/config": "^0.3.0",
+    "@rollup/plugin-commonjs": "^24.0.0",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^10.0.1",
+    "@types/chai": "^4.2.12",
+    "@types/mocha": "^10.0.0",
+    "@types/node": "^18.11.9",
+    "@types/require-from-string": "^1.2.0",
+    "@typescript-eslint/eslint-plugin": "^3.8.0",
+    "@typescript-eslint/parser": "^3.8.0",
+    "ajv-formats": "^3.0.0-rc.0",
+    "browserify": "^17.0.0",
     "chai": "^4.0.1",
-    "coveralls": "^3.0.1",
-    "del-cli": "^3.0.0",
-    "dot": "^1.0.3",
-    "eslint": "^7.3.1",
-    "gh-pages-generator": "^0.2.3",
-    "glob": "^7.0.0",
+    "cross-env": "^7.0.2",
+    "dayjs": "^1.10.4",
+    "dayjs-plugin-utc": "^0.1.2",
+    "eslint": "^7.8.1",
+    "eslint-config-prettier": "^7.0.0",
+    "fast-uri": "^2.1.0",
+    "glob": "^8.0.2",
+    "husky": "^8.0.2",
     "if-node-version": "^1.0.0",
+    "jimp": "^0.16.1",
     "js-beautify": "^1.7.3",
-    "jshint": "^2.10.2",
     "json-schema-test": "^2.0.0",
-    "karma": "^5.0.0",
+    "karma": "^6.0.0",
     "karma-chrome-launcher": "^3.0.0",
     "karma-mocha": "^2.0.0",
-    "karma-sauce-launcher": "^4.1.3",
-    "mocha": "^8.0.1",
+    "lint-staged": "^13.0.3",
+    "mocha": "^10.0.0",
+    "module-from-string": "^3.1.3",
+    "node-fetch": "^3.0.0",
     "nyc": "^15.0.0",
-    "pre-commit": "^1.1.1",
-    "require-globify": "^1.3.0",
-    "typescript": "^3.9.5",
-    "uglify-js": "^3.6.9",
-    "watch": "^1.0.0"
+    "prettier": "^2.3.1",
+    "re2": "^1.16.0",
+    "rollup": "^2.44.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "ts-node": "^10.0.0",
+    "tsify": "^5.0.2",
+    "typescript": "^4.8.0"
   },
   "collective": {
     "type": "opencollective",
