@@ -1,18 +1,15 @@
 {
   "name": "loader-utils",
-  "version": "2.0.4",
+  "version": "3.3.1",
   "author": "Tobias Koppers @sokra",
   "description": "utils for webpack loaders",
-  "dependencies": {
-    "big.js": "^5.2.2",
-    "emojis-list": "^3.0.0",
-    "json5": "^2.1.2"
-  },
+  "dependencies": {},
   "scripts": {
-    "lint": "eslint lib test",
+    "lint": "prettier --list-different . && eslint .",
     "pretest": "yarn lint",
     "test": "jest",
-    "test:ci": "jest --coverage",
+    "test:only": "jest --coverage",
+    "test:ci": "yarn test:only",
     "release": "yarn test && standard-version"
   },
   "license": "MIT",
@@ -21,16 +18,15 @@
     "url": "https://github.com/webpack/loader-utils.git"
   },
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">= 12.13.0"
   },
   "devDependencies": {
-    "coveralls": "^3.0.9",
-    "eslint": "^6.8.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "jest": "^25.1.0",
-    "prettier": "^1.19.1",
-    "standard-version": "^7.1.0"
+    "coveralls": "^3.1.1",
+    "eslint": "^8.0.1",
+    "eslint-plugin-node": "^11.1.0",
+    "jest": "^27.3.1",
+    "prettier": "^2.4.1",
+    "standard-version": "^9.3.2"
   },
   "main": "lib/index.js",
   "files": [
