@@ -1,7 +1,7 @@
 {
   "name": "postcss-js",
-  "version": "3.0.3",
-  "description": "PostCSS for React Inline Styles, Radium, Free Style and other CSS-in-JS",
+  "version": "4.0.1",
+  "description": "PostCSS for CSS-in-JS and styles in JS objects",
   "keywords": [
     "postcss",
     "postcss-runner",
@@ -15,21 +15,28 @@
   "license": "MIT",
   "repository": "postcss/postcss-js",
   "engines": {
-    "node": ">=10.0"
+    "node": "^12 || ^14 || >= 16"
   },
   "exports": {
     ".": {
       "require": "./index.js",
       "import": "./index.mjs"
     },
-    "./": "./"
+    "./package.json": "./package.json",
+    "./async": "./async.js",
+    "./objectifier": "./objectifier.js",
+    "./parser": "./parser.js",
+    "./process-result": "./process-result.js",
+    "./sync": "./sync.js"
   },
   "funding": {
     "type": "opencollective",
     "url": "https://opencollective.com/postcss/"
   },
+  "peerDependencies": {
+    "postcss": "^8.4.21"
+  },
   "dependencies": {
-    "camelcase-css": "^2.0.1",
-    "postcss": "^8.1.6"
+    "camelcase-css": "^2.0.1"
   }
 }
