@@ -12,96 +12,80 @@
     "tailwind": "lib/cli.js",
     "tailwindcss": "lib/cli.js"
   },
-  "contributors": [
-    "Adam Wathan <adam.wathan@gmail.com>",
-    "Jonathan Reinink <jonathan@reinink.ca>",
-    "David Hemphill <davidlee.hemphill@gmail.com>"
-  ],
   "scripts": {
-    "prebabelify": "npm run generate:plugin-list && rimraf lib",
-    "babelify": "babel src --out-dir lib --copy-files",
-    "postbabelify": "ncc build lib/cli-peer-dependencies.js -o peers",
-    "rebuild-fixtures": "npm run babelify && babel-node scripts/rebuildFixtures.js",
-    "prepublishOnly": "npm install --force && npm run babelify && babel-node scripts/build.js && node scripts/build-plugins.js",
+    "prebuild": "npm run generate && rimraf lib",
+    "build": "swc src --out-dir lib --copy-files",
+    "postbuild": "esbuild lib/cli-peer-dependencies.js --bundle --platform=node --outfile=peers/index.js --define:process.env.CSS_TRANSFORMER_WASM=false",
+    "rebuild-fixtures": "npm run build && node -r @swc/register scripts/rebuildFixtures.js",
     "style": "eslint .",
-    "test": "cross-env TAILWIND_MODE=build jest",
+    "pretest": "npm run generate",
+    "test": "jest",
     "test:integrations": "npm run test --prefix ./integrations",
     "install:integrations": "node scripts/install-integrations.js",
-    "posttest": "npm run style",
-    "compat": "node scripts/compat.js --prepare",
-    "compat:restore": "node scripts/compat.js --restore",
-    "generate:plugin-list": "babel-node scripts/create-plugin-list.js"
+    "generate:plugin-list": "node -r @swc/register scripts/create-plugin-list.js",
+    "generate:types": "node -r @swc/register scripts/generate-types.js",
+    "generate": "npm run generate:plugin-list && npm run generate:types",
+    "release-channel": "node ./scripts/release-channel.js",
+    "release-notes": "node ./scripts/release-notes.js",
+    "prepublishOnly": "npm install --force && npm run build"
   },
   "files": [
-    "dist/*.css",
+    "src/*",
     "cli/*",
     "lib/*",
     "peers/*",
     "scripts/*.js",
-    "stubs/*.stub.js",
+    "stubs/*",
     "nesting/*",
+    "types/**/*",
+    "*.d.ts",
     "*.css",
     "*.js"
   ],
   "devDependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.15.0",
-    "@babel/node": "^7.12.13",
-    "@babel/preset-env": "^7.15.0",
-    "@tailwindcss/aspect-ratio": "^0.2.1",
-    "@vercel/ncc": "^0.30.0",
-    "autoprefixer": "^10.3.3",
-    "babel-jest": "^27.0.6",
-    "clean-css": "5.1.4",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.0.8",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "jest": "^27.0.6",
-    "jest-diff": "^27.0.2",
-    "postcss": "^8.3.6",
-    "postcss-cli": "^8.3.1",
-    "prettier": "^2.3.2",
-    "rimraf": "^3.0.0"
-  },
-  "peerDependencies": {
-    "autoprefixer": "^10.0.2",
-    "postcss": "^8.0.9"
+    "@swc/cli": "0.1.62",
+    "@swc/core": "1.3.55",
+    "@swc/jest": "0.2.26",
+    "@swc/register": "0.1.10",
+    "autoprefixer": "^10.4.20",
+    "browserslist": "^4.24.0",
+    "concurrently": "^8.2.2",
+    "cssnano": "^6.1.2",
+    "esbuild": "^0.24.0",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.7.0",
+    "jest-diff": "^29.7.0",
+    "lightningcss": "1.27.0",
+    "prettier": "^2.8.8",
+    "rimraf": "^5.0.10",
+    "source-map-js": "^1.2.1",
+    "turbo": "^1.13.4"
   },
   "dependencies": {
-    "arg": "^5.0.1",
-    "bytes": "^3.0.0",
-    "chalk": "^4.1.2",
-    "chokidar": "^3.5.2",
-    "color": "^4.0.1",
-    "cosmiconfig": "^7.0.1",
-    "detective": "^5.2.0",
+    "@alloc/quick-lru": "^5.2.0",
+    "arg": "^5.0.2",
+    "chokidar": "^3.6.0",
     "didyoumean": "^1.2.2",
     "dlv": "^1.1.3",
-    "fast-glob": "^3.2.7",
-    "fs-extra": "^10.0.0",
-    "glob-parent": "^6.0.1",
-    "html-tags": "^3.1.0",
-    "is-color-stop": "^1.1.0",
-    "is-glob": "^4.0.1",
-    "lodash": "^4.17.21",
-    "lodash.topath": "^4.5.2",
-    "modern-normalize": "^1.1.0",
-    "node-emoji": "^1.11.0",
+    "fast-glob": "^3.3.2",
+    "glob-parent": "^6.0.2",
+    "is-glob": "^4.0.3",
+    "jiti": "^1.21.6",
+    "lilconfig": "^3.1.3",
+    "micromatch": "^4.0.8",
     "normalize-path": "^3.0.0",
-    "object-hash": "^2.2.0",
-    "postcss-js": "^3.0.3",
-    "postcss-load-config": "^3.1.0",
-    "postcss-nested": "5.0.6",
-    "postcss-selector-parser": "^6.0.6",
-    "postcss-value-parser": "^4.1.0",
-    "pretty-hrtime": "^1.0.3",
-    "purgecss": "^4.0.3",
-    "quick-lru": "^5.1.1",
-    "reduce-css-calc": "^2.1.8",
-    "resolve": "^1.20.0",
-    "tmp": "^0.2.1"
+    "object-hash": "^3.0.0",
+    "picocolors": "^1.1.1",
+    "postcss": "^8.4.47",
+    "postcss-import": "^15.1.0",
+    "postcss-js": "^4.0.1",
+    "postcss-load-config": "^4.0.2",
+    "postcss-nested": "^6.2.0",
+    "postcss-selector-parser": "^6.1.2",
+    "resolve": "^1.22.8",
+    "sucrase": "^3.35.0"
   },
   "browserslist": [
     "> 1%",
@@ -128,10 +100,19 @@
     ],
     "testPathIgnorePatterns": [
       "/node_modules/",
-      "/integrations/"
-    ]
+      "/integrations/",
+      "/standalone-cli/",
+      "\\.test\\.skip\\.js$"
+    ],
+    "transformIgnorePatterns": [
+      "node_modules/(?!lightningcss)"
+    ],
+    "transform": {
+      "\\.js$": "@swc/jest",
+      "\\.ts$": "@swc/jest"
+    }
   },
   "engines": {
-    "node": ">=12.13.0"
+    "node": ">=14.0.0"
   }
 }
