@@ -1,33 +1,9 @@
 import React, { useState } from 'react';
-import { Link, useHistory } from 'react-router-dom';
-import { Menu, X, User, Bell, LogOut } from 'lucide-react';
-import { useAuth } from '../../hooks/useAuth';
+import { Link, useNavigate } from 'react-router-dom';
+import { Menu, X } from 'lucide-react';
 
 const Navbar = () => {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [isProfileOpen, setIsProfileOpen] = useState(false);
-  const { user, logout, isAuthenticated } = useAuth();
-  const history = useHistory();
-  
-  const handleLogout = async () => {
-    await logout();
-    history.push('/');
-  };
-  
-  const getDashboardLink = () => {
-    if (!user) return '/';
-    
-    switch (user.role) {
-      case 'admin':
-        return '/admin/dashboard';
-      case 'employer':
-        return '/employer/dashboard';
-      case 'candidate':
-        return '/candidate/dashboard';
-      default:
-        return '/';
-    }
-  };
   
   return (
     <nav className="bg-white shadow-sm border-b border-gray-200">
@@ -54,65 +30,20 @@ const Navbar = () => {
           
           {/* User authentication section */}
           <div className="flex items-center">
-            {isAuthenticated ? (
-              <div className="hidden md:flex items-center">
-                <div className="ml-3 relative">
-                  <div>
-                    <button
-                      onClick={() => setIsProfileOpen(!isProfileOpen)}
-                      className="flex text-sm rounded-full focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
-                    >
-                      <span className="sr-only">Open user menu</span>
-                      <div className="h-8 w-8 rounded-full bg-blue-500 flex items-center justify-center text-white">
-                        {user?.name?.charAt(0) || user?.email?.charAt(0) || 'U'}
-                      </div>
-                    </button>
-                  </div>
-                  
-                  {isProfileOpen && (
-                    <div className="origin-top-right absolute right-0 mt-2 w-48 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5">
-                      <div className="py-1">
-                        <div className="px-4 py-2 text-sm text-gray-700 border-b">
-                          <p className="font-medium">{user?.name || 'User'}</p>
-                          <p className="text-xs text-gray-500">{user?.email}</p>
-                          <p className="text-xs text-gray-500 capitalize">{user?.role}</p>
-                        </div>
-                        <Link 
-                          to={getDashboardLink()}
-                          className="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                          onClick={() => setIsProfileOpen(false)}
-                        >
-                          <User className="mr-2 h-4 w-4" />
-                          Dashboard
-                        </Link>
-                        <button
-                          onClick={handleLogout}
-                          className="flex w-full items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                        >
-                          <LogOut className="mr-2 h-4 w-4" />
-                          Sign out
-                        </button>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </div>
-            ) : (
-              <div className="hidden md:flex items-center space-x-4">
-                <Link
-                  to="/login"
-                  className="px-4 py-2 text-sm font-medium text-gray-700 hover:text-gray-900"
-                >
-                  Log in
-                </Link>
-                <Link
-                  to="/register"
-                  className="px-4 py-2 text-sm font-medium text-white bg-blue-600 rounded-md hover:bg-blue-700"
-                >
-                  Sign up
-                </Link>
-              </div>
-            )}
+            <div className="hidden md:flex items-center space-x-4">
+              <Link
+                to="/login"
+                className="px-4 py-2 text-sm font-medium text-gray-700 hover:text-gray-900"
+              >
+                Log in
+              </Link>
+              <Link
+                to="/register"
+                className="px-4 py-2 text-sm font-medium text-white bg-blue-600 rounded-md hover:bg-blue-700"
+              >
+                Sign up
+              </Link>
+            </div>
             
             {/* Mobile menu button */}
             <div className="flex md:hidden">
@@ -156,52 +87,22 @@ const Navbar = () => {
             </Link>
           </div>
           
-          {isAuthenticated ? (
-            <div className="pt-4 pb-3 border-t border-gray-200">
-              <div className="flex items-center px-4">
-                <div className="flex-shrink-0">
-                  <div className="h-10 w-10 rounded-full bg-blue-500 flex items-center justify-center text-white">
-                    {user?.name?.charAt(0) || user?.email?.charAt(0) || 'U'}
-                  </div>
-                </div>
-                <div className="ml-3">
-                  <div className="text-base font-medium text-gray-800">{user?.name || 'User'}</div>
-                  <div className="text-sm font-medium text-gray-500">{user?.email}</div>
-                </div>
-              </div>
-              <div className="mt-3 space-y-1">
-                <Link
-                  to={getDashboardLink()}
-                  className="block px-4 py-2 text-base font-medium text-gray-700 hover:text-gray-900 hover:bg-gray-50"
-                >
-                  Dashboard
-                </Link>
-                <button
-                  onClick={handleLogout}
-                  className="block w-full text-left px-4 py-2 text-base font-medium text-gray-700 hover:text-gray-900 hover:bg-gray-50"
-                >
-                  Sign out
-                </button>
-              </div>
+          <div className="pt-4 pb-3 border-t border-gray-200">
+            <div className="flex items-center justify-center space-x-4">
+              <Link
+                to="/login"
+                className="px-4 py-2 text-base font-medium text-gray-700 hover:text-gray-900"
+              >
+                Log in
+              </Link>
+              <Link
+                to="/register"
+                className="px-4 py-2 text-base font-medium text-white bg-blue-600 rounded-md hover:bg-blue-700"
+              >
+                Sign up
+              </Link>
             </div>
-          ) : (
-            <div className="pt-4 pb-3 border-t border-gray-200">
-              <div className="flex items-center justify-center space-x-4">
-                <Link
-                  to="/login"
-                  className="px-4 py-2 text-base font-medium text-gray-700 hover:text-gray-900"
-                >
-                  Log in
-                </Link>
-                <Link
-                  to="/register"
-                  className="px-4 py-2 text-base font-medium text-white bg-blue-600 rounded-md hover:bg-blue-700"
-                >
-                  Sign up
-                </Link>
-              </div>
-            </div>
-          )}
+          </div>
         </div>
       )}
     </nav>
